+++ gazebo/sensors/Sensor.cc
@@ -174,8 +174,17 @@ void Sensor::Update(const bool _force)
   {
     if (this->useStrictRate)
     {
+      // rendering sensors (IMAGE category) has its own mechanism
+      // for throttling and lockstepping with physics. So throttle just
+      // physics sensors
+      if (this->dataPtr->category != IMAGE && !this->NeedsUpdate() && !_force)
+        return;
+
       if (this->UpdateImpl(_force))
+      {
+        this->lastUpdateTime = this->world->SimTime();;
         this->dataPtr->updated();
+      }
     }
     else
     {